requirements.txt
simplejson==3.11.1
python_dateutil==2.6.1
pandas==0.20.3
